test: split apply policy tests
