+++ test/clawsweeper.test.ts
@@ -1,6 +1,5 @@
 import assert from "node:assert/strict";
 import { execFileSync } from "node:child_process";
-import { createHash } from "node:crypto";
 import {
   chmodSync,
   existsSync,
@@ -53,7 +52,6 @@ import {
   renderReviewCommentFromReport,
   renderReviewContextBudgetForTest,
   rootCauseClusterFromReportForTest,
-  renderWorkPlanFromReport,
   restoreTreeModesForTest,
   reviewContextLedgerForTest,
   reviewDecisionSchemaText,
@@ -81,7 +79,21 @@ import {
 } from "../dist/commit-sweeper.js";
 import { parseArgs as parseClawsweeperArgs } from "../dist/clawsweeper-args.js";
 import { AUTOMATION_LIMITS } from "../dist/limits.js";
-import { closeDecision, git, item, tmpPrefix, withMockGh } from "./helpers.ts";
+import {
+  closeDecision,
+  git,
+  implementedCloseReport,
+  item,
+  lowSignalCloseReport,
+  markedReviewCommentForTest,
+  promotionGhMock,
+  reportWithSyncedReviewComment,
+  runApplyDecisionsForTest,
+  tmpPrefix,
+  withMockCodexProof,
+  withMockGh,
+  workPlanCandidateReport,
+} from "./helpers.ts";
 
 test("review prompt assets match tracked files", () => {
   assert.equal(reviewPromptTemplate(), readFileSync("prompts/review-item.md", "utf8"));
@@ -130,61 +142,6 @@ ${Object.entries(values)
   .map(([key, value]) => `${key}: ${value}`)
   .join("\n")}
 ---
-`;
-}
-
-function realBehaviorProofReportSection(overrides = {}) {
-  const values = {
-    status: "sufficient",
-    evidenceKind: "terminal",
-    needsContributorAction: false,
-    summary:
-      "The PR includes a terminal transcript from a real OpenClaw setup showing the fixed behavior after the patch.",
-    ...overrides,
-  };
-  return `## Real Behavior Proof
-
-Status: ${values.status}
-
-Evidence kind: ${values.evidenceKind}
-
-Needs contributor action: ${values.needsContributorAction}
-
-Summary: ${values.summary}
-`;
-}
-
-function prRatingReportSection(overrides = {}) {
-  const values = {
-    overallTier: "B",
-    proofTier: "A",
-    patchTier: "B",
-    overallLabel: "🐚 platinum hermit",
-    proofLabel: "🦞 diamond lobster",
-    patchLabel: "🐚 platinum hermit",
-    summary: "This PR has strong proof and normal merge-ready implementation quality.",
-    nextSteps: "- none",
-    ...overrides,
-  };
-  return `## PR Rating
-
-Overall tier: ${values.overallTier}
-
-Proof tier: ${values.proofTier}
-
-Patch tier: ${values.patchTier}
-
-Overall label: ${values.overallLabel}
-
-Proof label: ${values.proofLabel}
-
-Patch label: ${values.patchLabel}
-
-Summary: ${values.summary}
-
-Next rank-up steps:
-
-${values.nextSteps}
 `;
 }
 
@@ -872,63 +829,6 @@ test("ClawSweeper policy allows self implemented-on-main issue and PR close prop
   assert.equal(implementedIssue.ok, true);
 });
 
-function workPlanCandidateReport(overrides = {}) {
-  const frontmatter = {
-    number: 321,
-    repository: "openclaw/clawsweeper",
-    type: "issue",
-    title: "Render work plans",
-    reviewed_at: new Date().toISOString(),
-    review_status: "complete",
-    local_checkout_access: "verified",
-    decision: "keep_open",
-    action_taken: "kept_open",
-    work_candidate: "queue_fix_pr",
-    work_status: "candidate",
-    work_priority: "medium",
-    work_confidence: "high",
-    work_likely_files: JSON.stringify(["src/clawsweeper.ts", "test/clawsweeper.test.ts"]),
-    work_validation: JSON.stringify(["pnpm run check"]),
-    work_cluster_refs: JSON.stringify(["openclaw/clawsweeper#26"]),
-    ...overrides,
-  };
-  return `---
-${Object.entries(frontmatter)
-  .map(([key, value]) => `${key}: ${value}`)
-  .join("\n")}
----
-
-# #321: Render work plans
-
-## Summary
-
-The dashboard has queue_fix_pr candidates but no generated coding plan.
-
-## Repair Work Prompt
-
-Render generated plan markdown from existing report fields.
-`;
-}
-
-function implementedCloseReport(overrides = {}) {
-  return `${workPlanCandidateReport({
-    decision: "close",
-    action_taken: "proposed_close",
-    close_reason: "implemented_on_main",
-    confidence: "high",
-    work_candidate: "none",
-    work_status: "none",
-    item_snapshot_hash: "reviewed-snapshot",
-    item_created_at: "2026-05-01T00:00:00Z",
-    item_updated_at: "2026-05-01T00:00:00Z",
-    reproduction_status: "reproduced",
-    reproduction_confidence: "high",
-    fixed_sha: "1234567890abcdef1234567890abcdef12345678",
-    fixed_at: "2026-05-01T02:00:00Z",
-    ...overrides,
-  })}\n\n## Evidence\n\n- **main fix:** git show confirms current main has the replacement implementation and it is not in the latest release yet\n  - file: [src/clawsweeper.ts](https://github.com/openclaw/clawsweeper/blob/1234567890abcdef1234567890abcdef12345678/src/clawsweeper.ts)\n  - sha: [1234567890ab](https://github.com/openclaw/clawsweeper/commit/1234567890abcdef1234567890abcdef12345678)\n\n## Close Comment\n\nClosing this because the requested behavior is already on main.\n`;
-}
-
 function failedReviewReport(overrides = {}) {
   return `${workPlanCandidateReport({
     repository: "openclaw/openclaw",
@@ -1202,24 +1102,6 @@ process.exit(1);
   }
 });
 
-function lowSignalCloseReport(overrides = {}) {
-  return `${workPlanCandidateReport({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    decision: "close",
-    action_taken: "proposed_close",
-    close_reason: "low_signal_unmergeable_pr",
-    confidence: "high",
-    work_candidate: "none",
-    work_status: "none",
-    item_snapshot_hash: "reviewed-snapshot",
-    item_created_at: "2026-05-01T00:00:00Z",
-    item_updated_at: "2026-05-01T00:00:00Z",
-    author_association: "CONTRIBUTOR",
-    ...overrides,
-  })}\n\n## Evidence\n\n- **branch shape:** PR diff is mostly unrelated provider churn around a tiny possible useful tweak\n\n## Close Comment\n\nClosing this PR because the branch is not a useful landing base.\n`;
-}
-
 function unconfirmedProductDirectionCloseReport(overrides = {}) {
   return `${workPlanCandidateReport({
     repository: "openclaw/openclaw",
@@ -1390,1810 +1272,6 @@ function stripProofAndRatingFrontMatter(report: string): string {
     "",
   );
 }
-
-function promotionGhMock(options: {
-  number: number;
-  title?: string;
-  labels?: string[];
-  itemCreatedAt?: string;
-  itemUpdatedAt?: string;
-  itemUpdatedAtAfterLabelSync?: string;
-  itemUpdatedAtAfterLabelSyncLogPath?: string;
-  itemUpdatedAtAfterProof?: string;
-  itemUpdatedAtAfterProofLogPath?: string;
-  issueCommentCount?: number;
-  comment: string;
-  commentWriteLogPath?: string;
-  closeAppliedBodyLogPath?: string;
-  comments?: unknown[];
-  timeline?: unknown[];
-  linkedPulls?: Record<number, unknown>;
-  linkedPullsAfterProof?: Record<number, unknown>;
-  linkedIssues?: Record<number, unknown>;
-}) {
-  const title = options.title ?? "Stale F PR";
-  const itemCreatedAt = options.itemCreatedAt ?? "2026-02-01T00:00:00Z";
-  const itemUpdatedAt = options.itemUpdatedAt ?? "2026-05-01T00:00:00Z";
-  const comments = options.comments ?? [
-    {
-      id: 9000 + options.number,
-      html_url: `https://github.com/openclaw/openclaw/pull/${options.number}#issuecomment-${
-        9000 + options.number
-      }`,
-      created_at: "2026-05-01T01:00:00Z",
-      updated_at: "2026-05-01T01:00:00Z",
-      user: { login: "clawsweeper[bot]" },
-      body: options.comment,
-    },
-  ];
-  const issueCommentCount = options.issueCommentCount ?? comments.length;
-  const timeline = options.timeline ?? [];
-  const linkedPulls = options.linkedPulls ?? {};
-  const linkedIssues = options.linkedIssues ?? {};
-  return `
-	const { appendFileSync, existsSync } = require("fs");
-	const rawArgs = process.argv.slice(2);
-	const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-	const path = args[1] || "";
-	const slurp = args.includes("--slurp");
-	const jqIndex = args.indexOf("--jq");
-	const jq = jqIndex >= 0 ? args[jqIndex + 1] : "";
-	const comments = ${JSON.stringify(comments)};
-	const timeline = ${JSON.stringify(timeline)};
-	const linkedPulls = ${JSON.stringify(linkedPulls)};
-	const linkedPullsAfterProof = ${JSON.stringify(options.linkedPullsAfterProof ?? {})};
-	const linkedIssues = ${JSON.stringify(linkedIssues)};
-	const commentWriteLogPath = ${JSON.stringify(options.commentWriteLogPath ?? "")};
-	const closeAppliedBodyLogPath = ${JSON.stringify(options.closeAppliedBodyLogPath ?? "")};
-	const number = ${options.number};
-		const title = ${JSON.stringify(title)};
-		const labels = ${JSON.stringify(options.labels ?? ["status: 📣 needs proof"])};
-		const itemCreatedAt = ${JSON.stringify(itemCreatedAt)};
-		const itemUpdatedAt = ${JSON.stringify(itemUpdatedAt)};
-		const itemUpdatedAtAfterLabelSync = ${JSON.stringify(
-      options.itemUpdatedAtAfterLabelSync ?? "",
-    )};
-		const itemUpdatedAtAfterLabelSyncLogPath = ${JSON.stringify(
-      options.itemUpdatedAtAfterLabelSyncLogPath ?? "",
-    )};
-		const itemUpdatedAtAfterProof = ${JSON.stringify(options.itemUpdatedAtAfterProof ?? "")};
-		const itemUpdatedAtAfterProofLogPath = ${JSON.stringify(
-      options.itemUpdatedAtAfterProofLogPath ?? "",
-    )};
-		const proofHasRun = () =>
-		  itemUpdatedAtAfterProofLogPath &&
-		  existsSync(itemUpdatedAtAfterProofLogPath);
-		const liveLinkedPulls = proofHasRun()
-		  ? { ...linkedPulls, ...linkedPullsAfterProof }
-		  : linkedPulls;
-		const liveUpdatedAt =
-		  itemUpdatedAtAfterProof &&
-		  itemUpdatedAtAfterProofLogPath &&
-		  existsSync(itemUpdatedAtAfterProofLogPath)
-		    ? itemUpdatedAtAfterProof
-		    : itemUpdatedAtAfterLabelSync &&
-		        itemUpdatedAtAfterLabelSyncLogPath &&
-		        existsSync(itemUpdatedAtAfterLabelSyncLogPath)
-		      ? itemUpdatedAtAfterLabelSync
-		      : itemUpdatedAt;
-	const issueCommentCount = ${issueCommentCount};
-	if (args[0] === "api" && args[1] === "-i" && new RegExp("/issues/" + number + "/timeline(?:\\\\?|$)").test(args[2] || "")) {
-	  console.log("HTTP/2 200\\n\\n" + JSON.stringify(timeline));
-	} else if (args[0] === "api" && new RegExp("/issues/" + number + "/comments$").test(path) && args.includes("--method")) {
-	  if (commentWriteLogPath) appendFileSync(commentWriteLogPath, args.join(" ") + "\\n");
-	  if (closeAppliedBodyLogPath) {
-	    const input = args[args.indexOf("--input") + 1];
-	    appendFileSync(closeAppliedBodyLogPath, JSON.parse(require("fs").readFileSync(input, "utf8")).body + "\\n---body---\\n");
-	  }
-	  console.log("");
-	} else if (args[0] === "api" && new RegExp("/issues/comments/\\\\d+$").test(path) && args.includes("--method")) {
-	  if (commentWriteLogPath) appendFileSync(commentWriteLogPath, args.join(" ") + "\\n");
-	  console.log("");
-	} else if (args[0] === "api" && new RegExp("/issues/" + number + "/comments(?:\\\\?|$)").test(path)) {
-	  console.log(JSON.stringify(slurp ? [comments] : comments));
-	} else if (args[0] === "api" && new RegExp("/issues/" + number + "/timeline(?:\\\\?|$)").test(path)) {
-  console.log(JSON.stringify(slurp ? [timeline] : timeline));
-} else if (args[0] === "api" && new RegExp("/issues/" + number + "$").test(path)) {
-  console.log(JSON.stringify({
-    number,
-    title,
-    html_url: "https://github.com/openclaw/openclaw/pull/" + number,
-    body: "Stale PR body.",
-    created_at: itemCreatedAt,
-    updated_at: liveUpdatedAt,
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels,
-    comments: issueCommentCount,
-    pull_request: { url: "https://api.github.com/repos/openclaw/openclaw/pulls/" + number }
-  }));
-} else if (args[0] === "api" && new RegExp("/pulls/" + number + "$").test(path)) {
-  console.log(JSON.stringify({
-    number,
-    title,
-    html_url: "https://github.com/openclaw/openclaw/pull/" + number,
-    state: "open",
-    changed_files: 2,
-    commits: 1,
-    review_comments: 0,
-    body: "Stale PR body.",
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/openclaw" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/openclaw" } },
-    user: { login: "reporter" }
-  }));
-	} else if (args[0] === "api" && /\\/pulls\\/(\\d+)$/.test(path)) {
-	  const linkedNumber = Number((path.match(/\\/pulls\\/(\\d+)$/) || [])[1]);
-	  if (!liveLinkedPulls[linkedNumber]) {
-	    console.error("unexpected linked pull", linkedNumber);
-	    process.exit(1);
-	  }
-	  console.log(JSON.stringify(liveLinkedPulls[linkedNumber]));
-	} else if (args[0] === "api" && /\\/issues\\/(\\d+)\\/comments(?:\\?|$)/.test(path)) {
-	  const linkedNumber = Number((path.match(/\\/issues\\/(\\d+)\\/comments/) || [])[1]);
-	  const linkedIssue = liveLinkedPulls[linkedNumber] || linkedIssues[linkedNumber];
-  if (!linkedIssue) {
-    console.error("unexpected linked comments", linkedNumber);
-    process.exit(1);
-  }
-  if (linkedIssue.commentsError) {
-    console.error(linkedIssue.commentsError);
-    process.exit(1);
-  }
-  const linkedComments = Array.isArray(linkedIssue.comments)
-    ? linkedIssue.comments
-    : [];
-  console.log(JSON.stringify(slurp ? [linkedComments] : linkedComments));
-	} else if (args[0] === "api" && /\\/issues\\/(\\d+)$/.test(path)) {
-	  const linkedNumber = Number((path.match(/\\/issues\\/(\\d+)$/) || [])[1]);
-	  const linkedIssue = liveLinkedPulls[linkedNumber] || linkedIssues[linkedNumber];
-  if (!linkedIssue) {
-    console.error("unexpected linked issue", linkedNumber);
-    process.exit(1);
-  }
-  const labels = Array.isArray(linkedIssue.labels)
-    ? linkedIssue.labels.map((label) =>
-        typeof label === "string" ? label : label && label.name ? label.name : null,
-      ).filter(Boolean)
-    : [];
-  if (jq === "[.labels[].name]") {
-    console.log(JSON.stringify(labels));
-  } else {
-    console.log(JSON.stringify({
-      number: linkedNumber,
-      title: linkedIssue.title || ("PR #" + linkedNumber),
-      html_url: linkedIssue.html_url || ("https://github.com/openclaw/openclaw/pull/" + linkedNumber),
-      body: linkedIssue.body || "",
-      state: linkedIssue.state || "open",
-      labels: labels.map((name) => ({ name })),
-      comments: Array.isArray(linkedIssue.comments) ? linkedIssue.comments.length : 0,
-      pull_request: linkedIssue.pull_request || null,
-    }));
-  }
-	} else if (args[0] === "api" && /\\/pulls\\/(\\d+)\\/files(?:\\?|$)/.test(path)) {
-	  const linkedNumber = Number((path.match(/\\/pulls\\/(\\d+)\\/files/) || [])[1]);
-	  if (linkedNumber !== number && !liveLinkedPulls[linkedNumber]) {
-	    console.error("unexpected linked pull files", linkedNumber);
-	    process.exit(1);
-	  }
-	  const sourceFiles = [{ filename: "src/runtime.ts" }, { filename: "test/runtime.test.ts" }];
-	  const files = linkedNumber === number ? sourceFiles : liveLinkedPulls[linkedNumber].files || sourceFiles;
-  if (jq === "[.[].filename]") {
-    console.log(JSON.stringify(files.map((file) =>
-      typeof file === "string" ? file : file && file.filename ? file.filename : null,
-    ).filter(Boolean)));
-  } else {
-    console.log(JSON.stringify([files]));
-  }
-} else if (args[0] === "api" && new RegExp("/pulls/" + number + "/(files|commits|comments)(?:\\\\?|$)").test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "pr" && args[1] === "close" && args[2] === String(number)) {
-  console.log("");
-	} else if (args[0] === "issue" && args[1] === "edit") {
-	  if (itemUpdatedAtAfterLabelSyncLogPath) appendFileSync(itemUpdatedAtAfterLabelSyncLogPath, args.join(" ") + "\\n");
-	  console.log("");
-	} else if (args[0] === "label" || args[0] === "issue") {
-	  console.log("");
-	} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-}
-
-function markedReviewCommentForTest(number: number, body: string): string {
-  return `${body.trimEnd()}\n\n<!-- clawsweeper-review item=${number} -->`;
-}
-
-function sha256ForTest(text: string): string {
-  return createHash("sha256").update(text).digest("hex");
-}
-
-function reportWithSyncedReviewComment(
-  report: string,
-  number: number,
-  reason = "none",
-): {
-  report: string;
-  comment: string;
-} {
-  const comment = markedReviewCommentForTest(number, renderReviewCommentFromReport(report, reason));
-  return {
-    report: report.replace(
-      /^---\n/,
-      [
-        "---",
-        `review_comment_sha256: ${sha256ForTest(comment)}`,
-        `review_comment_id: ${9000 + number}`,
-        `review_comment_url: https://github.com/openclaw/clawsweeper/issues/${number}#issuecomment-${9000 + number}`,
-        "review_comment_synced_at: 2026-05-01T01:00:00Z",
-        "",
-      ].join("\n"),
-    ),
-    comment,
-  };
-}
-
-function withMockCodexProof(
-  root: string,
-  result:
-    | {
-        type: "decision";
-        decision: "covered" | "keep_open";
-        reason: string;
-        invocationLogPath?: string;
-        expectedPromptIncludes?: string;
-        unexpectedPromptIncludes?: string;
-        coveredPromptIncludes?: string;
-        keepOpenPromptIncludes?: string;
-      }
-    | { type: "failure"; message: string; invocationLogPath?: string },
-  run: () => void,
-): void {
-  const originalPath = process.env.PATH;
-  const binDir = join(root, "bin");
-  mkdirSync(binDir, { recursive: true });
-  const codexPath = join(binDir, "codex");
-  const script =
-    result.type === "decision"
-      ? `#!/usr/bin/env node
-	const { appendFileSync, writeFileSync } = require("fs");
-	const args = process.argv.slice(2);
-	const outputPath = args[args.indexOf("--output-last-message") + 1];
-	const invocationLogPath = ${JSON.stringify(result.invocationLogPath ?? "")};
-	const prompt = require("fs").readFileSync(0, "utf8");
-	const expectedPrompt = ${JSON.stringify(result.expectedPromptIncludes ?? "")};
-	if (expectedPrompt && !prompt.includes(expectedPrompt)) {
-	  process.stderr.write("missing expected proof prompt text: " + expectedPrompt);
-	  process.exit(1);
-	}
-	const unexpectedPrompt = ${JSON.stringify(result.unexpectedPromptIncludes ?? "")};
-		if (unexpectedPrompt && prompt.includes(unexpectedPrompt)) {
-		  process.stderr.write("unexpected proof prompt text: " + unexpectedPrompt);
-		  process.exit(1);
-		}
-		const coveredPrompt = ${JSON.stringify(result.coveredPromptIncludes ?? "")};
-		const keepOpenPrompt = ${JSON.stringify(result.keepOpenPromptIncludes ?? "")};
-		const decision = coveredPrompt && prompt.includes(coveredPrompt)
-		  ? "covered"
-		  : keepOpenPrompt && prompt.includes(keepOpenPrompt)
-		    ? "keep_open"
-		    : ${JSON.stringify(result.decision)};
-		if (invocationLogPath) appendFileSync(invocationLogPath, "proof\\n");
-		writeFileSync(outputPath, JSON.stringify({
-	  sourceSummary: "PR A updates the provider route.",
-	  coveringSummary: "PR B updates a different provider path.",
-	  coveredWork: decision === "covered" ? ["PR B includes PR A's provider route update."] : [],
-	  uniqueSourceWork: decision === "covered" ? [] : ["PR A's provider route update is still unique."],
-	  decision,
-	  reason: ${JSON.stringify(result.reason)}
-	}));
-`
-      : `#!/usr/bin/env node
-const { appendFileSync } = require("fs");
-const invocationLogPath = ${JSON.stringify(result.invocationLogPath ?? "")};
-if (invocationLogPath) appendFileSync(invocationLogPath, "proof\\n");
-console.error(${JSON.stringify(result.message)});
-process.exit(1);
-`;
-  writeFileSync(codexPath, script, { mode: 0o755 });
-  try {
-    process.env.PATH = `${binDir}:${originalPath ?? ""}`;
-    run();
-  } finally {
-    if (originalPath === undefined) delete process.env.PATH;
-    else process.env.PATH = originalPath;
-  }
-}
-
-function runApplyDecisionsForTest(options: {
-  targetRepo?: string;
-  itemsDir: string;
-  closedDir: string;
-  plansDir: string;
-  reportPath: string;
-  extraArgs?: string[];
-}): void {
-  execFileSync(process.execPath, [
-    "dist/clawsweeper.js",
-    "apply-decisions",
-    "--target-repo",
-    options.targetRepo ?? "openclaw/clawsweeper",
-    "--items-dir",
-    options.itemsDir,
-    "--closed-dir",
-    options.closedDir,
-    "--plans-dir",
-    options.plansDir,
-    "--report-path",
-    options.reportPath,
-    "--limit",
-    "10",
-    "--processed-limit",
-    "1",
-    "--close-delay-ms",
-    "0",
-    ...(options.extraArgs ?? []),
-  ]);
-}
-
-test("renderWorkPlanFromReport renders dashboard plan artifacts for fresh queue_fix_pr candidates", () => {
-  const plan = renderWorkPlanFromReport(workPlanCandidateReport(), {
-    reportPath: "records/openclaw-clawsweeper/items/321.md",
-  });
-  assert.ok(plan);
-  assert.match(plan, /# Coding Plan for openclaw\/clawsweeper#321: Render work plans/);
-  assert.match(plan, /Render generated plan markdown from existing report fields\./);
-  assert.match(plan, /- `src\/clawsweeper\.ts`/);
-  assert.match(plan, /- `pnpm run check`/);
-  assert.match(plan, /openclaw\/clawsweeper#26/);
-});
-
-test("renderWorkPlanFromReport returns null for stale, reclassified, or non-candidate reports", () => {
-  assert.equal(renderWorkPlanFromReport(workPlanCandidateReport({ work_candidate: "none" })), null);
-  assert.equal(
-    renderWorkPlanFromReport(workPlanCandidateReport({ work_status: "manual_review" })),
-    null,
-  );
-  assert.equal(renderWorkPlanFromReport(workPlanCandidateReport({ action_taken: "closed" })), null);
-  assert.equal(
-    renderWorkPlanFromReport(workPlanCandidateReport({ reviewed_at: "2026-01-01T00:00:00.000Z" })),
-    null,
-  );
-});
-
-test("apply-artifacts writes and removes generated work plans", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const artifactDir = join(root, "artifacts");
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    mkdirSync(artifactDir, { recursive: true });
-    writeFileSync(join(artifactDir, "321.md"), workPlanCandidateReport(), "utf8");
-    execFileSync(process.execPath, [
-      "dist/clawsweeper.js",
-      "apply-artifacts",
-      "--target-repo",
-      "openclaw/clawsweeper",
-      "--artifact-dir",
-      artifactDir,
-      "--items-dir",
-      itemsDir,
-      "--closed-dir",
-      closedDir,
-      "--plans-dir",
-      plansDir,
-      "--replay-closed-artifacts",
-      "--skip-reconcile",
-    ]);
-    const planPath = join(plansDir, "321.md");
-    assert.ok(existsSync(planPath));
-    assert.match(readFileSync(planPath, "utf8"), /## Plan\n\nRender generated plan markdown/);
-
-    writeFileSync(
-      join(artifactDir, "321.md"),
-      workPlanCandidateReport({ work_candidate: "none", work_status: "none" }),
-      "utf8",
-    );
-    execFileSync(process.execPath, [
-      "dist/clawsweeper.js",
-      "apply-artifacts",
-      "--target-repo",
-      "openclaw/clawsweeper",
-      "--artifact-dir",
-      artifactDir,
-      "--items-dir",
-      itemsDir,
-      "--closed-dir",
-      closedDir,
-      "--plans-dir",
-      plansDir,
-      "--replay-closed-artifacts",
-      "--skip-reconcile",
-    ]);
-    assert.equal(existsSync(planPath), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions removes archived work plans from the scoped plans directory", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const originalGhBin = process.env.GH_BIN;
-  const originalGhBinArgs = process.env.GH_BIN_ARGS;
-  const defaultPlanDir = join(process.cwd(), "records", "openclaw-clawsweeper", "plans");
-  const defaultPlanPath = join(defaultPlanDir, "321.md");
-  try {
-    const binDir = join(root, "bin");
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    mkdirSync(binDir, { recursive: true });
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    mkdirSync(defaultPlanDir, { recursive: true });
-    const ghMock = `#!/usr/bin/env node
-const args = process.argv.slice(2).join(" ");
-if (args.includes("/comments")) {
-  console.log(JSON.stringify([[]]));
-} else {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Render work plans",
-    html_url: "https://github.com/openclaw/clawsweeper/issues/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: "2026-05-02T00:00:00Z",
-    state: "closed",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    pull_request: null
-  }));
-}
-`;
-    writeFileSync(join(binDir, "gh.js"), ghMock, { mode: 0o755 });
-    writeFileSync(
-      join(itemsDir, "321.md"),
-      workPlanCandidateReport({
-        item_snapshot_hash: "reviewed-snapshot",
-        item_updated_at: "2026-05-01T00:00:00Z",
-      }),
-      "utf8",
-    );
-    writeFileSync(join(plansDir, "321.md"), "scoped generated plan\n", "utf8");
-    writeFileSync(defaultPlanPath, "default generated plan\n", "utf8");
-
-    process.env.GH_BIN = process.execPath;
-    process.env.GH_BIN_ARGS = JSON.stringify([join(binDir, "gh.js")]);
-    execFileSync(process.execPath, [
-      "dist/clawsweeper.js",
-      "apply-decisions",
-      "--target-repo",
-      "openclaw/clawsweeper",
-      "--items-dir",
-      itemsDir,
-      "--closed-dir",
-      closedDir,
-      "--plans-dir",
-      plansDir,
-      "--limit",
-      "1",
-      "--processed-limit",
-      "1",
-      "--close-delay-ms",
-      "0",
-    ]);
-
-    assert.equal(existsSync(join(plansDir, "321.md")), false);
-    assert.ok(existsSync(defaultPlanPath));
-    assert.ok(existsSync(join(closedDir, "321.md")));
-  } finally {
-    if (originalGhBin === undefined) delete process.env.GH_BIN;
-    else process.env.GH_BIN = originalGhBin;
-    if (originalGhBinArgs === undefined) delete process.env.GH_BIN_ARGS;
-    else process.env.GH_BIN_ARGS = originalGhBinArgs;
-    rmSync(root, { recursive: true, force: true });
-    rmSync(defaultPlanPath, { force: true });
-  }
-});
-
-test("apply-decisions skips advisory label sync when a close report changed since review", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    writeFileSync(
-      join(itemsDir, "321.md"),
-      workPlanCandidateReport({
-        decision: "close",
-        action_taken: "proposed_close",
-        close_reason: "implemented_on_main",
-        confidence: "high",
-        item_snapshot_hash: "reviewed-snapshot",
-        item_updated_at: "2026-05-01T00:00:00Z",
-        reproduction_status: "reproduced",
-        reproduction_confidence: "high",
-      }),
-      "utf8",
-    );
-
-    const ghMock = `
-const { appendFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Render work plans",
-    html_url: "https://github.com/openclaw/clawsweeper/issues/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-03T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    pull_request: null
-  }));
-} else if (args[0] === "issue" && args[1] === "view") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [] }));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({ itemsDir, closedDir, plansDir, reportPath });
-    });
-
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    assert.equal(
-      calls.some((args) => args[0] === "issue" && args[1] === "edit"),
-      false,
-    );
-    assert.equal(
-      calls.some((args) => args[0] === "label" && args[1] === "create"),
-      false,
-    );
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 321,
-        action: "skipped_changed_since_review",
-        reason: "updated_at changed",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions records PR label sync as ClawSweeper-owned churn", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    const itemPath = join(itemsDir, "74478.md");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    writeFileSync(
-      itemPath,
-      `${reportFrontMatter({
-        repository: "openclaw/clawsweeper",
-        type: "pull_request",
-        number: "74478",
-        title: "Record PR label churn",
-        url: "https://github.com/openclaw/clawsweeper/pull/74478",
-        decision: "keep_open",
-        close_reason: "none",
-        confidence: "high",
-        action_taken: "kept_open",
-        review_status: "complete",
-        local_checkout_access: "verified",
-        author: "contributor",
-        author_association: "CONTRIBUTOR",
-        labels: JSON.stringify([]),
-        item_category: "feature",
-        requires_new_feature: "true",
-        item_snapshot_hash: "snapshot-a",
-        item_updated_at: "2026-05-19T20:00:00Z",
-        pull_head_sha: "abc123def456",
-      })}
-
-## Summary
-
-This PR has complete review metadata and needs only ClawSweeper-owned labels.
-
-${realBehaviorProofReportSection({ evidenceKind: "screenshot" })}
-
-${prRatingReportSection({ overallTier: "A" })}
-
-## Feature Showcase
-
-Status: showcase
-
-Reason: This unlocks a notably useful maintainer workflow that did not exist before.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`,
-      "utf8",
-    );
-
-    const ghMock = `
-const { appendFileSync, readFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && /\\/issues\\/74478$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 74478,
-    title: "Record PR label churn",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/74478",
-    created_at: "2026-05-19T19:00:00Z",
-    updated_at: "2026-05-19T20:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "contributor" },
-    labels: [],
-    pull_request: {}
-  }));
-} else if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/74478\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/74478\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/pulls\\/74478$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 74478,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/74478",
-    state: "open",
-    changed_files: 1,
-    commits: 1,
-    review_comments: 0,
-    head: { sha: "abc123def456", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "contributor" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/74478\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/74478\\/comments(?:\\?|$)/.test(path)) {
-  if (args.includes("--method") && args.includes("POST")) {
-    const input = args[args.indexOf("--input") + 1];
-    appendFileSync(logPath, JSON.stringify(["posted-comment-body", JSON.parse(readFileSync(input, "utf8")).body]) + "\\n");
-    console.log(JSON.stringify({
-      id: 987478,
-      html_url: "https://github.com/openclaw/clawsweeper/pull/74478#issuecomment-987478"
-    }));
-  } else {
-    console.log(JSON.stringify([[]]));
-  }
-} else if (args[0] === "label" && args[1] === "create") {
-  console.log(JSON.stringify({ name: args[2] }));
-} else if (args[0] === "issue" && args[1] === "edit") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: ["--sync-comments-only", "--item-numbers", "74478"],
-      });
-    });
-
-    const report = readFileSync(itemPath, "utf8");
-    assert.match(report, /^labels_synced_at: /m);
-    assert.match(report, /proof: sufficient/);
-    assert.match(report, /proof: 📸 screenshot/);
-    assert.match(report, /rating: 🦞 diamond lobster/);
-    assert.match(report, /feature: ✨ showcase/);
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .map((line) => JSON.parse(line));
-    assert(
-      calls.some(
-        (args) => args[0] === "label" && args[1] === "create" && args[2] === "feature: ✨ showcase",
-      ),
-    );
-    assert(
-      calls.some(
-        (args) =>
-          args[0] === "issue" &&
-          args[1] === "edit" &&
-          args.includes("--add-label") &&
-          args.includes("feature: ✨ showcase"),
-      ),
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions routes parsed security owner acceptance to maintainer review", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const labelLogPath = join(root, "label-sync.log");
-    const itemPath = join(itemsDir, "74480.md");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-
-    const sourceReport = `${reportFrontMatter({
-      repository: "openclaw/openclaw",
-      type: "pull_request",
-      number: "74480",
-      title: "Route owner security acceptance",
-      url: "https://github.com/openclaw/openclaw/pull/74480",
-      decision: "keep_open",
-      close_reason: "none",
-      confidence: "high",
-      action_taken: "kept_open",
-      review_status: "complete",
-      local_checkout_access: "verified",
-      author: "contributor",
-      author_association: "CONTRIBUTOR",
-      labels: JSON.stringify(["status: ⏳ waiting on author"]),
-      item_snapshot_hash: "reviewed-snapshot",
-      item_created_at: "2026-02-01T00:00:00Z",
-      item_updated_at: "2026-05-01T00:00:00Z",
-      pull_head_sha: "head-sha",
-      merge_risk_options: JSON.stringify([
-        {
-          title: "Accept the reviewed security tradeoff",
-          body: "A maintainer may accept this bounded security tradeoff before merge.",
-          category: "accept_risk",
-          recommended: true,
-          automergeInstruction: "",
-        },
-      ]),
-    })}
-
-## Summary
-
-The patch is correct and the remaining security decision belongs to a maintainer.
-
-${realBehaviorProofReportSection()}
-
-${prRatingReportSection()}
-
-## Security Review
-
-Status: needs_attention
-
-Summary: A maintainer must explicitly accept the bounded security tradeoff.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.95
-
-Full review comments:
-
-- none
-`;
-    const synced = reportWithSyncedReviewComment(sourceReport, 74480);
-    writeFileSync(itemPath, synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 74480,
-        title: "Route owner security acceptance",
-        labels: ["status: ⏳ waiting on author"],
-        comment: synced.comment,
-        itemUpdatedAtAfterLabelSync: "2026-05-01T00:01:00Z",
-        itemUpdatedAtAfterLabelSyncLogPath: labelLogPath,
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          targetRepo: "openclaw/openclaw",
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: ["--sync-comments-only", "--item-numbers", "74480"],
-        });
-      },
-    );
-
-    const updatedReport = readFileSync(itemPath, "utf8");
-    assert.match(updatedReport, /status: 👀 ready for maintainer look/);
-    assert.doesNotMatch(updatedReport, /status: ⏳ waiting on author/);
-    const labelCalls = readFileSync(labelLogPath, "utf8");
-    assert.match(labelCalls, /--remove-label status: ⏳ waiting on author/);
-    assert.match(labelCalls, /--add-label status: 👀 ready for maintainer look/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions refreshes recent PR comments after label sync adds justifications", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    const itemPath = join(itemsDir, "74479.md");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    writeFileSync(
-      itemPath,
-      `${reportFrontMatter({
-        repository: "openclaw/clawsweeper",
-        type: "pull_request",
-        number: "74479",
-        title: "Refresh label explanation",
-        url: "https://github.com/openclaw/clawsweeper/pull/74479",
-        decision: "keep_open",
-        close_reason: "none",
-        confidence: "high",
-        action_taken: "kept_open",
-        review_status: "complete",
-        local_checkout_access: "verified",
-        author: "contributor",
-        author_association: "CONTRIBUTOR",
-        labels: JSON.stringify([]),
-        item_snapshot_hash: "snapshot-a",
-        item_updated_at: "2026-05-19T20:00:00Z",
-        pull_head_sha: "abc123def456",
-        review_comment_synced_at: "2026-05-19T23:59:00Z",
-      })}
-
-## Summary
-
-This PR needs labels and the latest comment must explain them.
-
-${realBehaviorProofReportSection({ evidenceKind: "screenshot" })}
-
-${prRatingReportSection({ overallTier: "A" })}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`,
-      "utf8",
-    );
-
-    const staleCommentBody =
-      "Codex review: needs maintainer review before merge.\n\n<!-- clawsweeper-review item=74479 -->";
-    const ghMock = `
-const { appendFileSync, readFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const staleCommentBody = ${JSON.stringify(staleCommentBody)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && /\\/issues\\/74479$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 74479,
-    title: "Refresh label explanation",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/74479",
-    created_at: "2026-05-19T19:00:00Z",
-    updated_at: "2026-05-19T20:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "contributor" },
-    labels: [],
-    pull_request: {}
-  }));
-} else if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/74479\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/74479\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/pulls\\/74479$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 74479,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/74479",
-    state: "open",
-    changed_files: 1,
-    commits: 1,
-    review_comments: 0,
-    head: { sha: "abc123def456", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "contributor" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/74479\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/74479\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[
-    {
-      id: 987479,
-      html_url: "https://github.com/openclaw/clawsweeper/pull/74479#issuecomment-987479",
-      body: staleCommentBody,
-      user: { login: "clawsweeper[bot]" },
-      created_at: "2026-05-19T23:59:00Z",
-      updated_at: "2026-05-19T23:59:00Z"
-    }
-  ]]));
-} else if (args[0] === "api" && /\\/issues\\/comments\\/987479$/.test(path)) {
-  const input = args[args.indexOf("--input") + 1];
-  appendFileSync(logPath, JSON.stringify(["patched-review-body", JSON.parse(readFileSync(input, "utf8")).body]) + "\\n");
-  console.log(JSON.stringify({
-    id: 987479,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/74479#issuecomment-987479"
-  }));
-} else if (args[0] === "label" && args[1] === "create") {
-  console.log(JSON.stringify({ name: args[2] }));
-} else if (args[0] === "issue" && args[1] === "edit") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: [
-          "--sync-comments-only",
-          "--comment-sync-min-age-days",
-          "7",
-          "--item-numbers",
-          "74479",
-        ],
-      });
-    });
-
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    const patchedBody = calls.find((args) => args[0] === "patched-review-body")?.[1] ?? "";
-    assert.match(patchedBody, /Label justifications:/);
-    assert.match(patchedBody, /`proof: sufficient`/);
-    assert.match(patchedBody, /`proof: 📸 screenshot`/);
-    assert.match(patchedBody, /`rating: 🦞 diamond lobster`/);
-    assert.match(readFileSync(itemPath, "utf8"), /^labels_synced_at: /m);
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 74479,
-        action: "review_comment_synced",
-        reason: "updated durable Codex review comment",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not advisory-label close proposals before close gates finish", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const closeReport = workPlanCandidateReport({
-      decision: "close",
-      action_taken: "proposed_close",
-      close_reason: "implemented_on_main",
-      confidence: "high",
-      item_snapshot_hash: "reviewed-snapshot",
-      item_updated_at: "2026-05-01T00:00:00Z",
-      reproduction_status: "reproduced",
-      reproduction_confidence: "high",
-    });
-    const synced = reportWithSyncedReviewComment(closeReport, 321, "implemented_on_main");
-    writeFileSync(join(itemsDir, "321.md"), synced.report, "utf8");
-
-    const ghMock = `
-const { appendFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const comment = ${JSON.stringify(synced.comment)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  const timeline = Array.from({ length: 100 }, (_, index) => ({ id: index + 1 }));
-  console.log('HTTP/2 200\\nlink: <https://api.github.com/repos/openclaw/clawsweeper/issues/321/timeline?per_page=100&page=2>; rel="last"\\n\\n' + JSON.stringify(timeline));
-} else if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[{
-    id: 9321,
-    html_url: "https://github.com/openclaw/clawsweeper/issues/321#issuecomment-9321",
-    created_at: "2026-05-01T01:00:00Z",
-    updated_at: "2026-05-01T01:00:00Z",
-    user: { login: "clawsweeper[bot]" },
-    body: comment
-  }]]));
-} else if (args[0] === "api" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Render work plans",
-    html_url: "https://github.com/openclaw/clawsweeper/issues/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    comments: 0,
-    pull_request: null
-  }));
-} else if (args[0] === "issue" && args[1] === "view") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [] }));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: ["--apply-close-reasons", "stale_insufficient_info"],
-      });
-    });
-
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    assert.equal(
-      calls.some((args) => args[0] === "issue" && args[1] === "edit"),
-      false,
-    );
-    assert.equal(
-      calls.some((args) => args[0] === "label" && args[1] === "create"),
-      false,
-    );
-    assert.equal(
-      calls.some(
-        (args) =>
-          args[0] === "api" &&
-          (args[1] ?? "").endsWith("/issues/321/timeline?per_page=100") &&
-          args.includes("--paginate"),
-      ),
-      true,
-    );
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 321,
-        action: "kept_open",
-        reason: "close reason implemented_on_main is not enabled for this apply run",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions posts an explicit close-time note before closing PR proposals", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    const postedBodiesPath = join(root, "posted-bodies.jsonl");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const closeReport = `${workPlanCandidateReport({
-      type: "pull_request",
-      decision: "close",
-      action_taken: "proposed_close",
-      close_reason: "implemented_on_main",
-      confidence: "high",
-      item_snapshot_hash: "reviewed-snapshot",
-      item_updated_at: "2026-05-01T00:00:00Z",
-      reproduction_status: "reproduced",
-      reproduction_confidence: "high",
-      fixed_pr_url: "https://github.com/openclaw/clawsweeper/pull/900",
-      fixed_pr_number: "900",
-      fixed_sha: "1234567890abcdef1234567890abcdef12345678",
-      fixed_at: "2026-05-01T02:00:00Z",
-    })}\n\n## Evidence\n\n- **main fix:** git show confirms current main has the replacement implementation and it is not in the latest release yet\n  - file: [src/clawsweeper.ts](https://github.com/openclaw/clawsweeper/blob/1234567890abcdef1234567890abcdef12345678/src/clawsweeper.ts)\n  - sha: [1234567890ab](https://github.com/openclaw/clawsweeper/commit/1234567890abcdef1234567890abcdef12345678)\n\n## Close Comment\n\nClosing this PR because the fix is already on main.\n`;
-    const synced = reportWithSyncedReviewComment(closeReport, 321, "implemented_on_main");
-    writeFileSync(join(itemsDir, "321.md"), synced.report, "utf8");
-
-    const ghMock = `
-const { appendFileSync, readFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const postedBodiesPath = ${JSON.stringify(postedBodiesPath)};
-const comment = ${JSON.stringify(synced.comment)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
-  if (args.includes("--method") && args.includes("POST")) {
-    const input = args[args.indexOf("--input") + 1];
-    const payload = JSON.parse(readFileSync(input, "utf8"));
-    appendFileSync(postedBodiesPath, JSON.stringify(payload.body) + "\\n");
-    console.log(JSON.stringify({ id: 9322, html_url: "https://github.com/openclaw/clawsweeper/pull/321#issuecomment-9322" }));
-  } else {
-    console.log(JSON.stringify([[{
-      id: 9321,
-      html_url: "https://github.com/openclaw/clawsweeper/pull/321#issuecomment-9321",
-      created_at: "2026-05-01T01:00:00Z",
-      updated_at: "2026-05-01T01:00:00Z",
-      user: { login: "clawsweeper[bot]" },
-      body: comment
-    }]]));
-  }
-} else if (args[0] === "api" && /\\/issues\\/comments\\/9321$/.test(path)) {
-  console.log(JSON.stringify({ id: 9321, html_url: "https://github.com/openclaw/clawsweeper/pull/321#issuecomment-9321" }));
-} else if (args[0] === "api" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Render work plans",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "MEMBER",
-    user: { login: "reporter" },
-    labels: ["maintainer"],
-    comments: 1,
-    pull_request: { url: "https://api.github.com/repos/openclaw/clawsweeper/pulls/321" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    state: "open",
-    changed_files: 0,
-    commits: 0,
-    review_comments: 0,
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "reporter" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/321\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "pr" && args[1] === "close" && args[2] === "321") {
-  console.log("");
-} else if (args[0] === "issue" && args[1] === "edit") {
-  console.log("");
-} else if (args[0] === "label") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: ["--apply-kind", "all", "--processed-limit", "2"],
-      });
-    });
-
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    const postIndex = calls.findIndex(
-      (args) =>
-        args[0] === "api" &&
-        (args[1] ?? "").endsWith("/issues/321/comments") &&
-        args.includes("POST"),
-    );
-    const closeIndex = calls.findIndex(
-      (args) => args[0] === "pr" && args[1] === "close" && args[2] === "321",
-    );
-    assert.ok(postIndex >= 0);
-    assert.ok(closeIndex > postIndex);
-    const postedBodies = readFileSync(postedBodiesPath, "utf8")
-      .trim()
-      .split("\n")
-      .map((line) => JSON.parse(line) as string);
-    assert.equal(postedBodies.length, 1);
-    assert.match(postedBodies[0], /ClawSweeper applied the proposed close for this PR/);
-    assert.match(postedBodies[0], /Close reason: already implemented on main/);
-    assert.match(postedBodies[0], /durable ClawSweeper review/);
-    assert.match(postedBodies[0], /clawsweeper-close-applied item=321/);
-    assert.ok(existsSync(join(closedDir, "321.md")));
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 321,
-        action: "review_comment_synced",
-        reason: "updated durable Codex review comment",
-      },
-      {
-        number: 321,
-        action: "closed",
-        reason: "already implemented on main; posted close-applied comment",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions keeps low-signal PRs open when live maintainer comments exist", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const logPath = join(root, "gh.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const closeReport = lowSignalCloseReport({ number: 322, title: "Add provider clamp" });
-    const synced = reportWithSyncedReviewComment(closeReport, 322, "low_signal_unmergeable_pr");
-    writeFileSync(join(itemsDir, "322.md"), synced.report, "utf8");
-
-    const ghMock = `
-const { appendFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const comment = ${JSON.stringify(synced.comment)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/322\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/322\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[
-    {
-      id: 9322,
-      html_url: "https://github.com/openclaw/clawsweeper/pull/322#issuecomment-9322",
-      created_at: "2026-05-01T01:00:00Z",
-      updated_at: "2026-05-01T01:00:00Z",
-      author_association: "NONE",
-      user: { login: "clawsweeper[bot]" },
-      body: comment
-    },
-    {
-      id: 9323,
-      html_url: "https://github.com/openclaw/clawsweeper/pull/322#issuecomment-9323",
-      created_at: "2026-05-01T01:30:00Z",
-      updated_at: "2026-05-01T01:30:00Z",
-      author_association: "MEMBER",
-      user: { login: "maintainer" },
-      body: "I am taking a look."
-    }
-  ]]));
-} else if (args[0] === "api" && /\\/issues\\/322\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/322$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 322,
-    title: "Add provider clamp",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/322",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    assignees: [],
-    comments: 2,
-    pull_request: { url: "https://api.github.com/repos/openclaw/clawsweeper/pulls/322" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/322$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 322,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/322",
-    state: "open",
-    changed_files: 4,
-    commits: 1,
-    review_comments: 0,
-    requested_reviewers: [],
-    requested_teams: [],
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "reporter" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/322\\/reviews(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/pulls\\/322\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "issue" && args[1] === "edit") {
-  console.log("");
-} else if (args[0] === "label") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        targetRepo: "openclaw/openclaw",
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: [
-          "--apply-kind",
-          "all",
-          "--processed-limit",
-          "2",
-          "--apply-close-reasons",
-          "low_signal_unmergeable_pr",
-        ],
-      });
-    });
-
-    const calls = readFileSync(logPath, "utf8")
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as string[]);
-    assert.equal(
-      calls.some((args) => args[0] === "pr" && args[1] === "close"),
-      false,
-    );
-    assert.equal(existsSync(join(closedDir, "322.md")), false);
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 322,
-        action: "kept_open",
-        reason: "maintainer issue comment blocks low-signal auto-close",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions retries legacy fixed close skips", () => {
-  for (const actionTaken of ["skipped_maintainer_authored", "skipped_invalid_decision"]) {
-    const root = mkdtempSync(tmpPrefix);
-    try {
-      const itemsDir = join(root, "items");
-      const closedDir = join(root, "closed");
-      const plansDir = join(root, "plans");
-      const reportPath = join(root, "apply-report.json");
-      const logPath = join(root, "gh.log");
-      mkdirSync(itemsDir, { recursive: true });
-      mkdirSync(plansDir, { recursive: true });
-      const closeReport = implementedCloseReport({
-        type: "pull_request",
-        action_taken: actionTaken,
-        author_association: "MEMBER",
-        labels: JSON.stringify(["maintainer"]),
-      }).replace(
-        "## Close Comment\n\nClosing this because the requested behavior is already on main.\n",
-        "## Close Comment\n\n_No close comment posted._\n",
-      );
-      const synced = reportWithSyncedReviewComment(closeReport, 321, "implemented_on_main");
-      writeFileSync(join(itemsDir, "321.md"), synced.report, "utf8");
-
-      const ghMock = `
-const { appendFileSync } = require("fs");
-const logPath = ${JSON.stringify(logPath)};
-const comment = ${JSON.stringify(synced.comment)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-appendFileSync(logPath, JSON.stringify(args) + "\\n");
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[{
-    id: 9321,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321#issuecomment-9321",
-    created_at: "2026-05-01T01:00:00Z",
-    updated_at: "2026-05-01T01:00:00Z",
-    user: { login: "clawsweeper[bot]" },
-    body: comment
-  }]]));
-} else if (args[0] === "api" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  if (args.includes("--method") && args.includes("PATCH")) {
-    console.log(JSON.stringify({ state: "closed" }));
-  } else {
-    console.log(JSON.stringify({
-      number: 321,
-      title: "Render work plans",
-      html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-      created_at: "2026-05-01T00:00:00Z",
-      updated_at: "2026-05-01T00:00:00Z",
-      closed_at: null,
-      state: "open",
-      locked: false,
-      active_lock_reason: null,
-      author_association: "MEMBER",
-      user: { login: "maintainer" },
-      labels: ["maintainer"],
-      comments: 1,
-      pull_request: { url: "https://api.github.com/repos/openclaw/clawsweeper/pulls/321" }
-    }));
-  }
-} else if (args[0] === "issue" && args[1] === "view") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [] }));
-} else if (args[0] === "api" && /\\/pulls\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    state: "open",
-    changed_files: 0,
-    commits: 0,
-    review_comments: 0,
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "maintainer" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/321\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-      withMockGh(root, ghMock, () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: ["--dry-run", "--apply-kind", "all", "--processed-limit", "2"],
-        });
-      });
-
-      assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-        {
-          number: 321,
-          action: "review_comment_synced",
-          reason: "would update durable Codex review comment",
-        },
-        {
-          number: 321,
-          action: "closed",
-          reason:
-            "dry-run: would close as already implemented on main; dry-run: would post close-applied comment",
-        },
-      ]);
-    } finally {
-      rmSync(root, { recursive: true, force: true });
-    }
-  }
-});
-
-test("apply-decisions retries legacy kept-open close reports", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const closeReport = implementedCloseReport({
-      type: "pull_request",
-      action_taken: "kept_open",
-    });
-    const synced = reportWithSyncedReviewComment(closeReport, 321, "implemented_on_main");
-    writeFileSync(join(itemsDir, "321.md"), synced.report, "utf8");
-
-    const ghMock = `
-const comment = ${JSON.stringify(synced.comment)};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[{
-    id: 9321,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321#issuecomment-9321",
-    created_at: "2026-05-01T01:00:00Z",
-    updated_at: "2026-05-01T01:00:00Z",
-    user: { login: "clawsweeper[bot]" },
-    body: comment
-  }]]));
-} else if (args[0] === "api" && /\\/issues\\/321\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Render work plans",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    comments: 1,
-    pull_request: { url: "https://api.github.com/repos/openclaw/clawsweeper/pulls/321" }
-  }));
-} else if (args[0] === "issue" && args[1] === "view") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [] }));
-} else if (args[0] === "api" && /\\/pulls\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    state: "open",
-    changed_files: 0,
-    commits: 0,
-    review_comments: 0,
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "reporter" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/321\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: ["--dry-run", "--apply-kind", "all", "--processed-limit", "2"],
-      });
-    });
-
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 321,
-        action: "review_comment_synced",
-        reason: "would update durable Codex review comment",
-      },
-      {
-        number: 321,
-        action: "closed",
-        reason:
-          "dry-run: would close as already implemented on main; dry-run: would post close-applied comment",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions pair-closes issues blocked by closeable linked PRs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const issueSynced = reportWithSyncedReviewComment(
-      implementedCloseReport({
-        repository: "openclaw/openclaw",
-        number: 320,
-        type: "issue",
-        title: "Issue fixed by main",
-        action_taken: "skipped_open_closing_pr",
-      }),
-      320,
-      "implemented_on_main",
-    );
-    const pullSynced = reportWithSyncedReviewComment(
-      implementedCloseReport({
-        repository: "openclaw/openclaw",
-        number: 321,
-        type: "pull_request",
-        title: "Obsolete linked PR",
-        action_taken: "kept_open",
-      }),
-      321,
-      "implemented_on_main",
-    );
-    writeFileSync(join(itemsDir, "320.md"), issueSynced.report, "utf8");
-    writeFileSync(join(itemsDir, "321.md"), pullSynced.report, "utf8");
-
-    const ghMock = `
-const comments = {
-  320: ${JSON.stringify(issueSynced.comment)},
-  321: ${JSON.stringify(pullSynced.comment)}
-};
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-const path = args[1] || "";
-const issueNumber = (path.match(/\\/issues\\/(\\d+)/) || [])[1];
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/(320|321)\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/(320|321)\\/comments(?:\\?|$)/.test(path)) {
-  const number = Number(issueNumber);
-  console.log(JSON.stringify([[{
-    id: 9000 + number,
-    html_url: "https://github.com/openclaw/clawsweeper/issues/" + number + "#issuecomment-" + (9000 + number),
-    created_at: "2026-05-01T01:00:00Z",
-    updated_at: "2026-05-01T01:00:00Z",
-    user: { login: "clawsweeper[bot]" },
-    body: comments[number]
-  }]]));
-} else if (args[0] === "api" && /\\/issues\\/(320|321)\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/320$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 320,
-    title: "Issue fixed by main",
-    html_url: "https://github.com/openclaw/clawsweeper/issues/320",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    comments: 1,
-    pull_request: null
-  }));
-} else if (args[0] === "api" && /\\/issues\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Obsolete linked PR",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    comments: 1,
-    pull_request: { url: "https://api.github.com/repos/openclaw/clawsweeper/pulls/321" }
-  }));
-} else if (args[0] === "issue" && args[1] === "view" && args[2] === "320") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [{ number: 321 }] }));
-} else if (args[0] === "issue" && args[1] === "view") {
-  console.log(JSON.stringify({ closedByPullRequestsReferences: [] }));
-} else if (args[0] === "api" && /\\/pulls\\/321$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 321,
-    title: "Obsolete linked PR",
-    html_url: "https://github.com/openclaw/clawsweeper/pull/321",
-    state: "open",
-    changed_files: 0,
-    commits: 0,
-    review_comments: 0,
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/clawsweeper" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/clawsweeper" } },
-    user: { login: "reporter" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/321\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      runApplyDecisionsForTest({
-        itemsDir,
-        closedDir,
-        plansDir,
-        reportPath,
-        extraArgs: [
-          "--target-repo",
-          "openclaw/openclaw",
-          "--dry-run",
-          "--apply-kind",
-          "all",
-          "--processed-limit",
-          "4",
-        ],
-      });
-    });
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.number === 320 && entry.action === "closed"),
-      true,
-    );
-    assert.equal(
-      report.some((entry) => entry.number === 321 && entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
 
 test("apply-decisions upgrades live no-diff kept-open PRs to duplicate closes", () => {
   const root = mkdtempSync(tmpPrefix);
